Getting much closer to Vulkan render

+++ content/shaders/basic.vert
@@ -7,7 +7,7 @@ layout (location = 1) in vec3 Normal;
 layout (location = 2) in vec4 Tangent;
 layout (location = 3) in vec2 UV0;
 layout (location = 4) in vec4 Color;
-layout (location = 5) in vec4 JointIds;
+layout (location = 5) in ivec4 JointIds;
 layout (location = 6) in vec4 JointWeights;
 layout (location = 7) in vec3 MorphTarget;
